Move js to ts (#30605)
(cherry picked from commit 6ac382e34ee8c1c766bc2d48e061cde50d6b9d83)
