+++ packages/devextreme/js/__internal/core/m_dom_adapter.ts
@@ -31,6 +31,7 @@ export interface DomAdapter {
   getRootNode: (element: HTMLElement) => Document | DocumentFragment;
   getAttribute: (element, name) => any;
   setAttribute: (element, name, value) => void;
+  createAttribute: (text: string, context?: Document) => Attr;
   removeAttribute: (element, name) => void;
   isElementNode: (element: any) => boolean;
   createElement: (tagName: string, context?: Document) => HTMLElement;
@@ -88,8 +89,10 @@ const nativeDOMAdapterStrategy: DomAdapter = {
     return context.createTextNode(text);
   },
 
-  createAttribute(text, context) {
+  createAttribute(text: string, context?: Document) {
+    // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
     context = context || this._document;
+    // @ts-expect-error
     return context.createAttribute(text);
   },
