+++ packages/devextreme/js/__internal/ui/m_themes_callback.ts
@@ -1,0 +1,3 @@
+import Callbacks from '@js/core/utils/callbacks';
+
+export const themeReadyCallback = Callbacks();
