+++ packages/devextreme/js/core/dom_adapter.d.ts
@@ -21,6 +21,7 @@ export interface DomAdapter {
   getRootNode: (element: HTMLElement) => Document | DocumentFragment;
   getAttribute: (element, name) => any;
   setAttribute: (element, name, value) => void;
+  createAttribute: (text: string, context?: Document) => Attr;
   removeAttribute: (element, name) => void;
   isElementNode: (element: any) => boolean;
   createElement: (tagName: string, context?: Document) => HTMLElement;
